backend/requirements.txt
flask
flask-cors
pymupdf
google-generativeai
python-dotenv
stripe
pyjwt
cryptography
requests
gunicorn
